requirements.txt
docx2txt~=0.8
requests~=2.31.0
streamlit~=1.32.0
beautifulsoup4~=4.12.3
python-dotenv~=1.0.1